backend/requirements.txt
appdirs==1.4.3
astroid==2.0.2
attrs==18.1.0
beautifulsoup4==4.6.1
black==18.6b4
click==6.7
Django==2.1
django-extensions==2.1.0
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.8.2
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
mongoengine==0.15.3
numpy==1.15.0
pandas==0.23.3
pylint==2.1.0
pymongo==3.7.1
python-dateutil==2.7.3
pytz==2018.5
six==1.11.0
toml==0.9.4
typed-ast==1.1.0
typing==3.6.4
wrapt==1.10.11
git+https://github.com/pyarun/django-mongoengine.git@ck-1#egg=django_mongoengine